pyproject.toml
[tool.poetry]
name = "byu-pathway-fb-ads-insights"
version = "1.0.0"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
requests = "^2.26.0"
google-cloud-bigquery = "^2.27.1"
google-cloud-tasks = "^2.5.2"
compose = "^1.2.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.9b0"
mypy = "^0.910"
types-requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-s --durations=0"

[tool.mypy]
python_version = 3.9
ignore_missing_imports = true
namespace_packages = true
